add relationships includes and check owners before edit or delete

+++ node/controllers/product.controller.js
@@ -5,7 +5,7 @@ const Category = Models.category;
 
 module.exports = {
     async get(req, res) {
-        const product = await Product.findByPk(req.params.product);
+        const product = await Product.findByPk(req.params.product, { include: [User, Location] });
         if (!product) {
             return res.status(404).send({ error: `Product with id ${req.params.product} not found` });
         }
@@ -44,6 +44,10 @@ module.exports = {
             return res.status(404).send({ error: `Product with id ${req.params.product} not found` });
         }
 
+        if (product.userId !== req.currentUser) {
+            return res.status(404).send({ error: `You are not the owner of this product` });
+        }
+
         const validation = Product.validate(req.body);
         if (validation.response !== 200) {
             return res.status(validation.response).send({ error: validation.error });
@@ -61,7 +65,7 @@ module.exports = {
 
         Product.update(validation.product, { where: { id: req.params.product } })
             .then(() => {
-                Product.findByPk(req.params.product)
+                Product.findByPk(req.params.product, { include: [User, Location] })
                     .then(product => {
                         return res.status(200).send(product)
                     })
@@ -70,18 +74,27 @@ module.exports = {
                 return res.status(400).send({ error: err });
             });
     },
-    delete(req, res) {
+    async delete(req, res) {
+        const product = await Product.findByPk(req.params.product, { include: [User, Location] });
+        if (!product) {
+            return res.status(404).send({ error: `Product with id ${req.params.product} not found` });
+        }
+        
+        if (product.userId !== req.currentUser) {
+            return res.status(404).send({ error: `You are not the owner of this product` });
+        }
+
         return Product.destroy({ where: { id: req.params.product } })
             .then(product => res.status(200).send({ deleted: product }))
             .catch(error => res.status(400).send({ error: error }))
     },
     findByName(req, res) {
-        return Product.findAll({ where: { name: { $like: `%${req.params.product}%` } } })
+        return Product.findAll({ include: [User, Location], where: { name: { $like: `%${req.params.product}%` } } })
             .then(product => res.status(200).send(product))
             .catch(error => res.status(400).send(error))
     },
     findAll(_, res) {
-        return Product.findAll({})
+        return Product.findAll({ include: [User, Location] })
             .then(product => res.status(200).send(product))
             .catch(error => res.status(400).send(error))
     }
