+++ node/controllers/rental.controller.js
@@ -7,7 +7,7 @@ const LocationController = require('./location.controller');
 
 module.exports = {
     async get(req, res) {
-        const rental = await Rental.findByPk(req.params.rental);
+        const rental = await Rental.findByPk(req.params.rental, { include: [User, Location] });
         if (!rental) {
             return res.status(404).send({ error: `Rental ${req.params.rental} not found` });
         }
@@ -45,6 +45,10 @@ module.exports = {
             return res.status(404).send({ error: `Rental with id ${req.params.rental} not found` });
         }
 
+        if (rental.userId !== req.currentUser) {
+            return res.status(404).send({ error: `You are not the owner of this rental` });
+        }
+
         const validation = Rental.validate(req.body);
         if (validation.response !== 200) {
             return res.status(validation.response).send({ error: validation.error });
@@ -52,7 +56,7 @@ module.exports = {
 
         Rental.update(req.body, { where: { id: rental.id } })
             .then(async () => {
-                Rental.findByPk(req.params.rental)
+                Rental.findByPk(req.params.rental, { include: [User, Location] })
                     .then(async rental => {
                         const location = await LocationController.createIfNotExist(validation.rental.location);
                         if (location) {
@@ -67,18 +71,27 @@ module.exports = {
                 return res.status(400).send({ error: err });
             });
     },
-    delete(req, res) {
+    async delete(req, res) {
+        const rental = await Rental.findByPk(req.params.rental, { include: [User, Location] });
+        if (!rental) {
+            return res.status(404).send({ error: `Rental ${req.params.rental} not found` });
+        }
+        
+        if (rental.userId !== req.currentUser) {
+            return res.status(404).send({ error: `You are not the owner of this rental` });
+        }
+
         return Rental.destroy({ where: { id: req.params.rental } })
             .then(rental => res.status(200).send({ deleted: rental }))
             .catch(error => res.status(400).send({ error: error }))
     },
     findByName(req, res) {
-        return Rental.findAll({ where: { name: { $like: `%${req.params.rental}%` } } })
+        return Rental.findAll({ include: [User, Location], where: { name: { $like: `%${req.params.rental}%` } } })
             .then(rentals => res.status(200).send(rentals))
             .catch(error => res.status(400).send(error))
     },
     findAll(_, res) {
-        return Rental.findAll({})
+        return Rental.findAll({ include: [User, Location] })
             .then(rentals => res.status(200).send(rentals))
             .catch(error => res.status(400).send(error))
     }
